da3/910B/requirements.txt
# official requirements
pre-commit
einops
huggingface_hub
imageio
numpy<2
opencv-python
open3d
fastapi
uvicorn
requests
typer
pillow
omegaconf
pillow_heif
safetensors



# 910B
conda create -n da3 --clone PyTorch-2.1.0
pip install moviepy==1.0.3 plyfile  trimesh evo e3nn fastapi uvicorn
 